% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_clustered_score.R
\name{cal_clustered_score}
\alias{cal_clustered_score}
\title{calculate clustered score
The function to calculate clustered score in noncoding regions}
\usage{
cal_clustered_score(list_result)
}
\arguments{
\item{list_result}{list format for the combined result}
}
\value{
None
}
\description{
calculate clustered score
The function to calculate clustered score in noncoding regions
}
