% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_mut_prob.R
\name{clustered_mut_prob}
\alias{clustered_mut_prob}
\title{combine mutations of inner and outer
The function to combine the inner mutations and outer mutations for each recurrent mutation}
\usage{
clustered_mut_prob(inner_pos_pro, outer_pos_pro, inner_mut, outer_mut,
  list_result)
}
\arguments{
\item{lr_result}{list format for the combined result}

\item{inner_pos_anno}{the file contains mutations and annotation values surrounded the recurrent position ±5 base pairs}

\item{outer_pos_anno:}{the file contains mutations and annotation values surrounded the recurrent position ±500 base pairs}

\item{inner_pos_pro:}{the file contains mutations surrounded the recurrent position ±5 base pairs and their mutation probilities}

\item{outer_pos_pro:}{the file contains mutations surrounded the recurrent position ±500 base pairs and their mutation probilities}
}
\value{
None
}
\description{
combine mutations of inner and outer
The function to combine the inner mutations and outer mutations for each recurrent mutation
}
