% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flank_mutation_prob.R
\name{flank_mutation_prob}
\alias{flank_mutation_prob}
\title{predict mutation probility
The function to predict mutation probility for each mutation}
\usage{
flank_mutation_prob(lr_result, inner_pos_anno, outer_pos_anno, inner_pos_prob,
  outer_pos_prob)
}
\arguments{
\item{lr_result}{trainning result used to predict mutation probability for each genomic position}

\item{inner_pos_anno}{the file contains mutations and annotation values surrounded the recurrent position ±5 base pairs}

\item{outer_pos_anno:}{the file contains mutations and annotation values surrounded the recurrent position ±500 base pairs}

\item{inner_pos_prob:}{output inner position result}

\item{outer_pos_prob:}{output outer position result}
}
\value{
None
}
\description{
predict mutation probility
The function to predict mutation probility for each mutation
}
