% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr_model.R
\name{get_lr_model}
\alias{get_lr_model}
\title{Training logistic regression model 
The function to train logistic regression model of annotated genomic regions}
\usage{
get_lr_model(pos_annotation, pos_mut, lr_result)
}
\arguments{
\item{pos_annotation}{the file contains genomic base pair features: base pair(bp), replicating time(reptime), TFBS signal(tfbs), GERP RS score(gerp), DNaseI signal(dnase), CpG Island(cpg), transcript status(gencode_anno)}

\item{pos_mut}{the file contains somatic mutations(single base substitution) in ICGC database}

\item{lr_result}{trainning result used to predict mutation probability for each genomic position}
}
\value{
None
}
\description{
Training logistic regression model 
The function to train logistic regression model of annotated genomic regions
}
